Code cleanup with README
